+++ src/gdb/breakpoints.h
@@ -9,8 +9,8 @@ unsigned int gdb_get_opcode(unsigned int addr);
 void gdb_reinstall_opcodes();
 void gdb_uninstall_opcodes();
 #else
-void gdb_insert_breakpoint(unsigned long addr, int len, int type);
-void gdb_remove_breakpoint(unsigned long addr, int len, int type);
+void gdb_insert_breakpoint(unsigned long long int addr, int len, int type);
+void gdb_remove_breakpoint(unsigned long long int addr, int len, int type);
 void gdb_remove_all();
 unsigned int gdb_get_opcode(unsigned long addr);
 void gdb_reinstall_opcodes();
